# delivery/http/api/docs/swagger.yaml
definitions:
  authentication_dto.AuthTokenDTO:
    properties:
      token:
        type: string
    type: object
...
  responses.ErrorResult:
    properties:
      code:
        type: integer
      error: {}
    type: object
  responses.InfoResult:
    properties:
      message:
        type: string
    type: object
...
  user_dto.UserDTO:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  user_payloads.UserDeletePayload:
    properties:
      id:
        type: integer
    type: object
  user_payloads.UserLoginPayload:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  user_payloads.UserProfileUpdatePayload:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  user_payloads.UserRegisterPayload:
    properties:
      age:
        type: integer
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: rocky@lazycats.id
    name: API Support
    url: https://rocky.my.id/
  description: This is a REST API for MyGram.
  title: MyGram
  version: "1.0"
paths:
...
  /users:
    delete:
      consumes:
      - application/json
      description: Deletes current user.
      parameters:
      - description: Delete User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user_payloads.UserDeletePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.InfoResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResult'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResult'
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_dto.UserDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResult'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResult'
      security:
      - ApiKeyAuth: []
      summary: Get current user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Updates current user.
      parameters:
      - description: Update User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user_payloads.UserProfileUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_dto.UserDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResult'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResult'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - user
  /users/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: User Login Request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/user_payloads.UserLoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authentication_dto.AuthTokenDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResult'
      summary: Login user
      tags:
      - user
  /users/register:
    post:
      consumes:
      - application/json
      description: Registers new user
      parameters:
      - description: User Register Request
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/user_payloads.UserRegisterPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authentication_dto.AuthTokenDTO'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResult'
      summary: Register new user
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"